requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
cssselect2==0.7.0
Django==4.0.5
django-ckeditor==6.4.2
django-filer==2.2.3
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
easy-thumbnails==2.8.3
lxml==4.9.1
mysqlclient==2.0.3
Pillow==9.1.1
reportlab==3.6.12
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.2.1
Unidecode==1.1.2
webencodings==0.5.1
